Example.User (monex v0.1.0)

Defines a User struct with custom equality (Eq) and ordering (Ord) implementations for functional, domain-driven comparisons.

In our domain, a User is identified primarily by their first_name and last_name. Custom comparison functions are also provided to compare users based on optional attributes, such as age.

Types

role()

@type role() :: :owner | :admin | :user

t()

@type t() :: %Example.User{
  age: Monex.Maybe.t(non_neg_integer()),
  first_name: String.t(),
  last_name: String.t(),
  role: role()
}

Functions

create_user(first, last, role, age \\ nil)

@spec create_user(String.t(), String.t(), role(), integer() | nil) :: t()

Creates a new User struct, accepting a first_name, last_name, role, and optional age.

If age is not provided, it defaults to Maybe.nothing().

Uses Maybe.lift_predicate/2 to wrap age as Just if it’s a valid integer, otherwise Nothing.

Examples

iex> Example.User.create_user("John", "Doe", :user, 30)
%Example.User{
  first_name: "John",
  last_name: "Doe",
  age: %Monex.Maybe.Just{value: 30},
  role: :user
}

iex> Example.User.create_user("Jane", "Smith", :admin)
%Example.User{
  first_name: "Jane",
  last_name: "Smith",
  age: %Monex.Maybe.Nothing{},
  role: :admin
}

eq_age()

Provides a custom Eq comparator for comparing Users by age.

This function uses contramap to lift equality to the age field of each user.

Examples

iex> eq_age = Example.User.eq_age()
iex> user1 = Example.User.create_user("Alice", "Smith", :user, 25)
iex> user2 = Example.User.create_user("Bob", "Brown", :admin, 25)
iex> Monex.Eq.Utils.eq?(user1, user2, eq_age)
true

ord_age()

Provides a custom Ord comparator for ordering Users by age.

This function uses contramap to lift ordering to the age field of each user.

Examples

iex> ord_age = Example.User.ord_age()
iex> user1 = Example.User.create_user("Alice", "Smith", :user, 20)
iex> user2 = Example.User.create_user("Bob", "Brown", :admin, 25)
iex> Monex.Ord.Utils.min(user1, user2, ord_age)
%Example.User{first_name: "Alice", last_name: "Smith", age: %Monex.Maybe.Just{value: 20}, role: :user}