Examples.Patron (monex v0.1.0)

The Examples.Patron module provides an example of how to represent a patron visiting a venue or amusement park. It shows how to use Monex’s Predicate module to combine and work with multiple conditions in an easy-to-understand way.

This module covers:

Creating a Patron with new/3.

Checking the patron’s height using too_short?/1, too_tall?/1, and valid_height?/1.

Checking and adjusting the number of tickets with has_ticket?/1, increment_ticket/1, and decrement_ticket/1.

You will also learn how to use predicates, a powerful concept in Monex, to validate multiple conditions about a patron.

Types

t()

@type t() :: %Examples.Patron{height: integer(), name: String.t(), tickets: integer()}

Functions

decrement_ticket(patron)

@spec decrement_ticket(t()) :: t()

Decrements the number of tickets a patron has by 1.

Examples

iex> patron = Examples.Patron.new("John", 170, 2)
iex> patron = Examples.Patron.decrement_ticket(patron)
iex> patron.tickets
1

has_ticket?(patron)

@spec has_ticket?(t()) :: boolean()

Checks if the patron has at least one ticket.

Examples

iex> patron_with_ticket = Examples.Patron.new("Ticket Holder", 180, 1)
iex> Examples.Patron.has_ticket?(patron_with_ticket)
true

iex> patron_no_ticket = Examples.Patron.new("No Ticket", 175, 0)
iex> Examples.Patron.has_ticket?(patron_no_ticket)
false

increment_ticket(patron)

@spec increment_ticket(t()) :: t()

Increments the number of tickets a patron has by 1.

Examples

iex> patron = Examples.Patron.new("John", 170, 2)
iex> patron = Examples.Patron.increment_ticket(patron)
iex> patron.tickets
3

new(name, height, tickets)

Creates a new Patron with the given name, height, and number of tickets.

Examples

iex> john = Examples.Patron.new("John", 170, 2)
%Examples.Patron{name: "John", height: 170, tickets: 2}

too_short?(patron)

@spec too_short?(t()) :: boolean()

Checks if the patron is too short (less than 150 cm).

Examples

iex> john = Examples.Patron.new("John", 140, 2)
iex> Examples.Patron.too_short?(john)
true

iex> jane = Examples.Patron.new("Jane", 160, 1)
iex> Examples.Patron.too_short?(jane)
false

too_tall?(patron)

@spec too_tall?(t()) :: boolean()

Checks if the patron is too tall (more than 200 cm).

Examples

iex> tall_patron = Examples.Patron.new("Tall Guy", 210, 1)
iex> Examples.Patron.too_tall?(tall_patron)
true

iex> average_patron = Examples.Patron.new("Average Joe", 170, 2)
iex> Examples.Patron.too_tall?(average_patron)
false

valid_height?(patron)

@spec valid_height?(t()) :: boolean()

Checks if the patron's height is valid, meaning they are neither too short nor too tall.

This combines the too_short?/1 and too_tall?/1 predicates using Monex's Predicate.p_and/2 and Predicate.p_not/1.

Examples

iex> short_patron = Examples.Patron.new("Shorty", 140, 1)
iex> Examples.Patron.valid_height?(short_patron)
false

iex> perfect_patron = Examples.Patron.new("Perfect Height", 180, 2)
iex> Examples.Patron.valid_height?(perfect_patron)
true