Monex.Identity (monex v0.1.0)

The Monex.Identity module represents the identity monad, where values are simply wrapped in a structure and operations are applied directly to those values.

This module implements the following protocols:

Monex.Monad: Implements the bind/2, map/2, and ap/2 functions for monadic operations.

Monex.Eq: Defines equality checks for Identity values.

Monex.Ord: Defines ordering logic for Identity values.

String.Chars: Converts an Identity value into a string representation.

The Identity monad provides a way to wrap a value in a minimal context and perform monadic operations.

Types

t(value)

@type t(value) :: %Monex.Identity{value: value}

Functions

extract(identity)

@spec extract(t(value)) :: value when value: term()

Extracts the value from an Identity.

Examples

iex> Monex.Identity.extract(Monex.Identity.pure(5))
5

get_eq(custom_eq)

get_ord(custom_ord)

pure(value)

@spec pure(value) :: t(value) when value: term()

Creates a new Identity value by wrapping a given value.

Examples

iex> Monex.Identity.pure(5)
%Monex.Identity{value: 5}