Monex.Ord protocol (monex v0.1.0)

The Monex.Ord protocol defines a set of comparison functions: lt?/2, le?/2, gt?/2, and ge?/2.

This protocol is intended for types that can be ordered, allowing values to be compared for their relative positions in a total order.

By implementing this protocol, you can provide custom logic for how values of a certain type are compared.

Fallback

The protocol uses @fallback_to_any true, which means if a specific type does not implement Monex.Ord, the default implementation for Any will be used, which relies on Elixir's built-in comparison operators.

Types

t()

@type t() :: term()

All the types that implement this protocol.

Functions

ge?(a, b)

Returns true if a is greater than or equal to b, otherwise returns false.

Examples

iex> Monex.Ord.ge?(Monex.Maybe.just(5), Monex.Maybe.just(3))
true

iex> Monex.Ord.ge?(Monex.Maybe.just(5), Monex.Maybe.just(5))
true

iex> Monex.Ord.ge?(Monex.Maybe.just(3), Monex.Maybe.just(5))
false

gt?(a, b)

Returns true if a is greater than b, otherwise returns false.

Examples

iex> Monex.Ord.gt?(Monex.Maybe.just(5), Monex.Maybe.just(3))
true

iex> Monex.Ord.gt?(Monex.Maybe.just(3), Monex.Maybe.just(5))
false

iex> Monex.Ord.gt?(Monex.Maybe.just(3), Monex.Maybe.nothing())
true

le?(a, b)

Returns true if a is less than or equal to b, otherwise returns false.

Examples

iex> Monex.Ord.le?(Monex.Maybe.just(3), Monex.Maybe.just(5))
true

iex> Monex.Ord.le?(Monex.Maybe.just(5), Monex.Maybe.just(5))
true

iex> Monex.Ord.le?(Monex.Maybe.just(5), Monex.Maybe.just(3))
false

lt?(a, b)

Returns true if a is less than b, otherwise returns false.

Examples

iex> Monex.Ord.lt?(Monex.Maybe.just(3), Monex.Maybe.just(5))
true

iex> Monex.Ord.lt?(Monex.Maybe.just(5), Monex.Maybe.just(3))
false

iex> Monex.Ord.lt?(Monex.Maybe.nothing(), Monex.Maybe.just(3))
true